requirements.txt
requests
rocketchat_API
ldap3
pyyaml